refactor(0000): share form budget hook for forecast sections
Extract duplicated budget/env form watching into useFormForecastBudget
to clear SonarQube new-code duplication.
