+++ app/components/public-cloud/sections/PublicCloudForecastSection.tsx
@@ -2,17 +2,10 @@
 
 import { Alert, Button } from '@mantine/core';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { useMemo } from 'react';
-import { useFormContext } from 'react-hook-form';
 import LoadingBox from '@/components/generic/LoadingBox';
-import {
-  buildRollingFiscalForecastMonths,
-  FISCAL_FORECAST_HORIZON_MONTHS,
-  getProviderBudgetCurrency,
-  sumEnabledEnvironmentBudgets,
-} from '@/components/public-cloud/forecast/forecast-grid-utils';
+import { FISCAL_FORECAST_HORIZON_MONTHS } from '@/components/public-cloud/forecast/forecast-grid-utils';
 import ProjectBudgetForecastPanel from '@/components/public-cloud/forecast/ProjectBudgetForecastPanel';
-import { useForecastBudgetCad } from '@/components/public-cloud/forecast/useForecastBudgetCad';
+import { useFormForecastBudget } from '@/components/public-cloud/forecast/useFormForecastBudget';
 import { getPublicCloudProductForecast } from '@/services/backend/public-cloud/forecast';
 import { usePublicProductState } from '@/states/global';
 
@@ -31,16 +24,7 @@ export default function PublicCloudForecastSection({ licencePlate }: Readonly<{ 
   const canViewForecast = Boolean(product?._permissions.viewForecast);
   const canEditForecast = Boolean(product?._permissions.editForecast);
   const queryClient = useQueryClient();
-  const { watch } = useFormContext();
-
-  const budgetDev = watch('budget.dev');
-  const budgetTest = watch('budget.test');
-  const budgetProd = watch('budget.prod');
-  const budgetTools = watch('budget.tools');
-  const envDev = watch('environmentsEnabled.development');
-  const envTest = watch('environmentsEnabled.test');
-  const envProd = watch('environmentsEnabled.production');
-  const envTools = watch('environmentsEnabled.tools');
+  const { budgetMonthlyTotal, budgetCurrency, draftMonthlyValues } = useFormForecastBudget(product?.provider);
 
   const { data, isLoading, isError, error, refetch } = useQuery({
     queryKey: ['forecast', licencePlate],
@@ -52,38 +36,6 @@ export default function PublicCloudForecastSection({ licencePlate }: Readonly<{ 
   const handleForecastSaved = () => queryClient.invalidateQueries({ queryKey: ['forecast', licencePlate] });
 
   const forecast = data?.forecast;
-
-  const budget = useMemo(
-    () => ({
-      dev: Number(budgetDev) || 0,
-      test: Number(budgetTest) || 0,
-      prod: Number(budgetProd) || 0,
-      tools: Number(budgetTools) || 0,
-    }),
-    [budgetDev, budgetTest, budgetProd, budgetTools],
-  );
-
-  const environmentsEnabled = useMemo(
-    () => ({
-      development: Boolean(envDev),
-      test: Boolean(envTest),
-      production: Boolean(envProd),
-      tools: Boolean(envTools),
-    }),
-    [envDev, envTest, envProd, envTools],
-  );
-
-  const budgetCurrency = getProviderBudgetCurrency(product?.provider);
-  const budgetMonthlyTotal = useMemo(
-    () => sumEnabledEnvironmentBudgets(budget, environmentsEnabled),
-    [budget, environmentsEnabled],
-  );
-  const { budgetMonthlyTotalCad } = useForecastBudgetCad(budgetMonthlyTotal, budgetCurrency);
-
-  const draftMonthlyValues = useMemo(
-    () => buildRollingFiscalForecastMonths(budgetMonthlyTotalCad ?? 0, 'CAD', new Date()),
-    [budgetMonthlyTotalCad],
-  );
 
   if (!product) return null;
   if (!canViewForecast) return null;
